Extract common db backend functions to reusable mixins
